requirements.txt
opencv-python==4.5.5.62
pillow==9.0.1
pytesseract==0.3.8
pyautogui==0.9.53
numpy==1.21.2
